+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.4.0
+
+* Add `__d`: it's `toString` method returns a regular translated string, not HTML
+* Add `I18n#concat(str)` method
+* Add `I18n#getDirect()`, which returns a `__d`-ified object
+
 ## 0.3.0 (2016-10-04)
 
 * Adapt plugin to alchemymvc 0.3.0